tests/wpt/web-platform-tests/html/cross-origin-opener-policy/access-reporting/openee-accessed_openee-coop-ro.https.html

<title>
  COOP reports are sent when the openee used COOP-RO+COEP and then its opener
  tries to access it.
</title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/get-host-info.sub.js></script>
<script src="/common/utils.js"></script>
<script src="./resources/dispatcher.js"></script>
<script>

const directory = "/html/cross-origin-opener-policy/access-reporting";
const executor_path = directory + "/resources/executor.html?pipe=";
const cross_origin = get_host_info().HTTPS_REMOTE_ORIGIN;
const coep_header = '|header(Cross-Origin-Embedder-Policy,require-corp)';

let operation = [
//[property      , operation                 ] ,
  ["blur"        , w => w.blur()             ] ,
  ["close"       , w => w.close()            ] ,
  ["closed"      , w => w.closed             ] ,
  ["focus"       , w => w.focus()            ] ,
  ["frames"      , w => w.frames             ] ,
  ["indexed"     , w => w[0]                 ] ,
  ["indexed"     , w => w[0] = ""            ] ,
  ["length"      , w => w.length             ] ,
  ["location"    , w => w.location           ] ,
  ["location"    , w => w.location = "#"     ] ,
  ["named"       , w => w["test"]            ] ,
  ["named"       , w => w["test"] = ""       ] ,
  ["opener"      , w => w.opener             ] ,
  ["opener"      , w => w.opener = ""        ] ,
  ["postMessage" , w => w.postMessage("")    ] ,
  ["postMessage" , w => w.postMessage("", "")] ,
  ["self"        , w => w.self               ] ,
  ["top"         , w => w.top                ] ,
  ["window"      , w => w.window             ] ,
];

operation.forEach(([property, op]) => {
  promise_test(async t => {
    const report_token = token();
    const executor_token = token();
    const callback_token = token();

    const reportTo = reportToHeaders(report_token);
    const openee_url = cross_origin + executor_path +
      reportTo.header + reportTo.coopReportOnlySameOriginHeader + coep_header +
      `&uuid=${executor_token}`;
    const openee = window.open(openee_url);
    t.add_cleanup(() => send(executor_token, "window.close()"))

    // 1. Make sure the new document to be loaded.
    send(executor_token, `
      send("${callback_token}", "Ready");
    `);
    let reply = await receive(callback_token);
    assert_equals(reply, "Ready");

    // 2. Try to access the openee. A report is sent, because of COOP-RO+COEP.
    try {op(openee)} catch(e) {}

    // 3. Check a report is sent to the openee.
    let report = await receiveReport(report_token, "access-to-coop-page")
    assert_not_equals(report, "timeout", "Report not received");
    assert_equals(report.type, "coop");
    assert_equals(report.url, openee_url.replace(/"/g, '%22'));
    assert_equals(report.body["violation-type"], "access-to-coop-page");
    assert_equals(report.body["disposition"], "reporting");
    assert_equals(report.body["effective-policy"], "same-origin-plus-coep");
    assert_equals(report.body["property"], property);
    assert_equals(report.body["source-file"], undefined);
    assert_equals(report.body["lineno"], undefined);
    assert_equals(report.body["colno"], undefined);
    // TODO(arthursonzogni): Add check for report > body > blocked-window-url
  }, `${op}`);
});

</script>
